refactor: format time
